movie-api/requirements.txt
boto3
requests
PyJWT

